set up condition to delete icon in relation to delete a post or a comment that is comparing user connected and original user id who created either the post or the comment

+++ backend/src/controllers/commentaires.js
@@ -2,7 +2,7 @@ const connection = require('../service/database');
 
 exports.createComment = (req, res) => {
     connection
-        .query("INSERT INTO commentaires (content_comment, user_id, post_id) VALUES (?,?,?)", [req.body.newCommentContent, req.body.user_id, req.params["postId"]])
+        .query("INSERT INTO commentaires (content_comment, user_id, post_id) VALUES (?,?,?)", [req.body.newCommentContent, req.body.userId, req.params["postId"]])
         .then((comment) => {
             res.status(201).json(comment);
         })
@@ -24,8 +24,9 @@ exports.modifyComment = (req, res) => {
 
 exports.deleteComment = (req, res) => {
     connection
-        .query("DELETE FROM commentaires WHERE id = ?", req.params['commentId'])
+        .query("DELETE FROM commentaires WHERE id = ? AND user_id = ?", [req.params['commentId'], req.body.userId])
         .then((commentDeleted) => {
+            console.log(commentDeleted)
             res.status(200).json(commentDeleted);
         })
         .catch((error) => {
